IS-1845: Refactor
